+++ extensions/integration-tests/src/tests/dacpac.test.ts
@@ -5,13 +5,13 @@
 
 import 'mocha';
 import * as azdata from 'azdata';
-import * as utils from '../utils';
+import * as utils from './utils';
 import * as path from 'path';
 import * as fs from 'fs';
 import * as os from 'os';
 import * as mssql from '../../../mssql';
 import * as vscode from 'vscode';
-import { getStandaloneServer } from '../testConfig';
+import { getStandaloneServer } from './testConfig';
 import * as assert from 'assert';
 import { promisify } from 'util';
