+++ extensions/integration-tests/src/tests/objectExplorer.test.ts
@@ -5,8 +5,8 @@
 
 import 'mocha';
 import * as azdata from 'azdata';
-import { getBdcServer, TestServerProfile, getAzureServer, getStandaloneServer } from '../testConfig';
-import { connectToServer, createDB, deleteDB, DefaultConnectTimeoutInMs, asyncTimeout } from '../utils';
+import { getBdcServer, TestServerProfile, getAzureServer, getStandaloneServer } from './testConfig';
+import { connectToServer, createDB, deleteDB, DefaultConnectTimeoutInMs, asyncTimeout } from './utils';
 import * as assert from 'assert';
 import { stressify } from 'adstest';
