+++ extensions/integration-tests/src/tests/testConfig.ts
